# Support larger cache line size for L2 cache