## Changes
Cache line is widened, with burst read/write supported within 1 cache line access.

## Motivation
Supporting a wider cache line would allow wishbone burst access with less frequent cache hit/miss checks.

## Throughput
We consider the following case:
- CPU makes a 64 bits * 8 cycles burst to the cache.
- Memory bus is 128 bits wide

The new cache would have the following geometry:
- Cache line is 64 bytes (exactly the access size)
- Burst access is aligned to cache lines

For now we assume the memory **cannot** buffer memory access.

|      |   Original   |      New     |
|:----:|:------------:|:------------:|
|  Hit |      15      |       8      |
| Miss | 31+4*WL+4*RL | 12+4*WL+4*RL |

With burst memory access implemented in #151. If we ignore overheads such as refresh, assuming WL=2 and RL=6.:
- The 4 back-to-back write can be condensed into a 6 cycles delay, before receiving the last ack.
- The 4 back-to-back read can be condensed into a 10 cycles delay, before receiving the last ack.

|      |   Original   |      New     | New (Mem Burst) |
|:----:|:------------:|:------------:|:---------------:|
|  Hit |      15      |       8      |        8        |
| Miss | 63 | 48 |         28        |

## Performance
Without modifications:
`Async throughput:   2.94MiB/s`
| Test                 | Mean (MiB/s) |  std (MiB/s) |
| -------------------- | ------------ | ------------ |
| I32 Array (1MB) H2D  |         2.41 |         0.02 |
| I32 Array (1MB) D2H  |         2.88 |         0.01 |
| I32 Array (1KB) H2D  |         0.51 |         0.04 |
| I32 Array (1KB) D2H  |         0.50 |         0.01 |
| Bytes List (1MB) H2D |         2.38 |         0.08 |
| Bytes List (1MB) D2H |         2.82 |         0.01 |
| Bytes List (1KB) H2D |         0.50 |         0.04 |
| Bytes List (1KB) D2H |         0.49 |         0.00 |
| Bytes (1MB) H2D      |         2.41 |         0.06 |
| Bytes (1MB) D2H      |         2.90 |         0.00 |
| Bytes (1KB) H2D      |         0.50 |         0.04 |
| Bytes (1KB) D2H      |         0.49 |         0.00 |
| I32 List (1MB) H2D   |         2.55 |         0.04 |
| I32 List (1MB) D2H   |         2.88 |         0.00 |
| I32 List (1KB) H2D   |         0.45 |         0.02 |
| I32 List (1KB) D2H   |         0.48 |         0.02 |

With #151 on ARTIQ Kasli v2, and L2 cache line size at 64 bytes.
`Async throughput:   3.19MiB/s`
| Test                 | Mean (MiB/s) |  std (MiB/s) |
| -------------------- | ------------ | ------------ |
| I32 Array (1MB) H2D  |         2.65 |         0.04 |
| I32 Array (1MB) D2H  |         3.14 |         0.00 |
| I32 Array (1KB) H2D  |         0.52 |         0.05 |
| I32 Array (1KB) D2H  |         0.52 |         0.07 |
| Bytes List (1MB) H2D |         2.86 |         0.02 |
| Bytes List (1MB) D2H |         3.06 |         0.01 |
| Bytes List (1KB) H2D |         0.53 |         0.05 |
| Bytes List (1KB) D2H |         0.56 |         0.06 |
| Bytes (1MB) H2D      |         2.73 |         0.06 |
| Bytes (1MB) D2H      |         3.16 |         0.00 |
| Bytes (1KB) H2D      |         0.53 |         0.03 |
| Bytes (1KB) D2H      |         0.54 |         0.01 |
| I32 List (1MB) H2D   |         2.75 |         0.03 |
| I32 List (1MB) D2H   |         3.13 |         0.02 |
| I32 List (1KB) H2D   |         0.52 |         0.02 |
| I32 List (1KB) D2H   |         0.54 |         0.00 |

## Compatibility
Setting `l2_line_size=1` (default) would yield the original cache geometry before patch, in which the cache line is exactly as wide as the slave bus width.